+++ CHANGELOG.md
@@ -75,7 +75,7 @@ All notable changes to this project will be documented in this file.
   - `toNegativeIntOrNull` and `toNegativeIntOrThrow`
   - `toStrictlyPositiveIntOrNull` and `toStrictlyPositiveIntOrThrow`
   - `toStrictlyNegativeIntOrNull` and `toStrictlyNegativeIntOrThrow`
-  - `toStrictlyPositiveDoubleOrNull`.
+  - `toStrictlyPositiveDoubleOrNull` and `toStrictlyPositiveDoubleOrThrow`.
 
 [#258]: https://github.com/kotools/types/issues/258